+++ lib/src/exceptions/path_dne.dart
@@ -1,0 +1,11 @@
+/// Thrown when attempting to perform an operation on a path in [AutoreleasingCache] that doesn't exist.
+class PathDoesNotExistException implements Exception {
+  /// The path that generated the exception
+  String offendingPath;
+
+  PathDoesNotExistException(this.offendingPath);
+
+  @override
+  String toString() =>
+      "The path '$offendingPath' does not exist yet access was attempted.";
+}